motion: check console task when available
motion_sense_init() checks if it was called in hooks task or console task
context. In Zephyr console task is not available so TASK_ID_CONSOLE is
not defined. Because of that HAS_TASK_CONSOLE is used to choose if check
for hooks and console task or only for hooks task.

BUG=none
TEST=makeall
BRANCH=none

Signed-off-by: Tomasz Michalec <tm@semihalf.com>
Change-Id: Ie51b043209d5228c074bbc9a693bf7b98cf2a598