battery_fuel_gauge: Check length of device string
In the common battery fuel gauge framework, the boards may specify
strings for the expected battery packs to be used at runtime.  Some
specified battery strings may be shorter than what is provided by the
fuel gauge as the later portion of the string can vary according to lot
code and/or manufacture date.  This commit simply checks up to the
length of the string provided by the board in order to find a match.

BUG=b:183875170
BRANCH=dedede
TEST=Build and flash a dedede device, verify that the battery can be
found.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ib1c4690ca2b676e6e019015c4dec1d57899479b1