chgstv2: Call board_check_extpower on AC change
Some boards don't have an interrupt which indicate that extpower is
present or not.  In some cases, the EC may miss an edge if the call
comes to early to check whether AC is ok.  However, the charger task
periodically polls this state.

This commit has the charger task call `board_check_extpower` when it
notices that the AC presence has changed.  This can ultimately end up
triggering the HOOK_AC_CHANGE hook.

BUG=b:189911786
BRANCH=dedede
TEST=Build and flash boten, plug in charger, suspend DUT, unplug DUT,
verify that DUT wakes up even after the first call to check AC OK
occurs.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Id0b60289cfb14509fd781942afbcedfa9e127a8a