LN9310: Disable auto-recovery options
Disable all auto-recovery options for the ln9310 so it will not
attempt an automatic restart in the event a protection circuit prevents
a normal startup.  The workaround in 2876277 adds an EC-controlled
segment to the ln9310 startup sequence which must be run for every
startup, so auto-recovery (which would only trigger the internal 9310
startup sequence and not the EC-controlled portion) should be
disabled.

BRANCH=Trogdor
BUG=b:185308433
TEST=Should not break the current boot flow, i.e. power-up and
power-down (using long-press of the power button) should function
similar to before this change.  This version needs to be compared
to the previous firmware version by booting some number of boards
with each firmware and confirming the old version has the issue
described in the bug and the new version does not. Ideally each
board should be booted repeatedly (100s or 1000s of cycles).

Change-Id: I4b9d032951204c9dfc9512f2dddc748b3aa8d9ef