zephyr: Ensure that CONFIG_USBC_PPC can be defined
At present the Kconfig for this option never makes it through to the
EC build. This does not seem to cause ill effects, but it does create
a different between the configs enabled in ECOS and Zephyr, when
CONFIG_PLATFORM_EC_USBC_PPC is defined. Fix it by adding a
suitable #define.

BUG=b:189855648
BRANCH=none
TEST=with other CLs, build asurada for Zephyr

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ifd2dd9ee63d1575effd99376f6ea5fbd1baac5a0