cret/raa489000: Control ASGATE state
There is 0.1V leakage from Charger to Vbus
when a charger is removed in S5.
Control ASGATE state to prevent the leakage.

This commit uses the charger side registers to
control the ASGATE when selecting our active charge port.
This is done in addition to the existing implementation
which uses the TCPCI registers to control ASGATE.

BUG=b:189299803
BRANCH=dedede
TEST=Build and flash Cret, the leakage has been improved
from 0.1V to 0.001V

Signed-off-by: johnwc_yeh <johnwc_yeh@compal.corp-partner.google.com>
Change-Id: Ib81ca651b0708149fae7b4df27f66107b47f8490