zephyr: shim: add mpu enable and disable support
Add support for allowing the EC code to control the MPU setup.

ECOS code used to handle configuration as well as MPU enable/disable.
Zephyr has its own way of handling MPU configuration, so that part of
the EC code is compiled out, instead make sure that the fixed MPU
regions (that cover code & data) are disabled before the EC code starts,
and then enable them when mpu_enable() is called

Zephyr APIs do not expose any memory protection funciontality outside of
the arch specific code, so this sets the MPU regiers directly. That's
not very elegant, but it's done in a sample application as well [1], so
it should be an acceptable approach.

Disabling the MPU unit entirely could also be an option, but only doing
the fixed regions has the advantage of leaving the stack protection and
null pointer detection regions enabled.

[1] https://github.com/zephyrproject-rtos/zephyr/blob/main/samples/application_development/code_relocation/src/main.c

BRANCH=none
BUG=b:180039888
TEST=actually tried to write in the static region, verified it fauls
after sysjump disable

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I020c347e5f1b929500d4cf1c4c400dcd67e78dfa