common/timer: Fallback to udelay when interrupt conditions are not met
To call usleep() we must be sure that interrupts are enabled and we are
not in interrupt context. This is required because hardware timer is
used to make sleeping task ready. Also on Cortex-M task switching is not
working properly when interrupts are disabled.

If above conditions are not met, just print warning and use udelay().

BUG=b:190597666
BRANCH=none
TEST=make -j buildall
TEST=Call usleep() when interrupts are disabled and make sure that
     warning is printed and EC doesn't crash. One can implement console
     command to test this.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Ia38ef0f0511ca1298d3153fa77e169019f4c4c31