gsctool: add --dauntless flag and default to haven
For commands that do not take in an image, we cannot tell if we are
communicating with a haven or dauntless chip automatically. Add a way to
specify if we are communicating with a dauntless chip. Default to haven
if the optional --dauntless flag is not present

BUG=b:190659004
TEST=execute `gsctool -I` on haven chip and see success
TEST=execute `gsctool -I --dauntless` on haven chip and see failure that
it cannot find device with dauntless VID:PID

Change-Id: Ia92422846d283eba5b27cfaa130c782f13af90ca
Signed-off-by: Jett Rink <jettrink@chromium.org>