asurada: enable HOST_SLEEP_STATE to smooth suspend trasnsition
MT8192 previously didn't enable the command due to it already has
AP_IN_SLEEP_L, but we found that if we don't have the command,
EC might have rare chance to interrupt AP with hostevent when AP
in transition to S3, and cause the suspend failure.

This CL enables CONFIG_POWER_TRACK_HOST_SLEEP_STATE so the EC is
able to handle the AP's hint to disable non-wakeup hostevent when
it's transisting to S3.

BUG=b:186709406
TEST=ensure EC received the AP's note
BRANCH=asurada

Change-Id: I7a2a7eefc7620817ca446fd24370375130daf6be
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>