build system: Link with LTO partitioning strategy "one"
By putting all object files in a single LTO partition there's less
room to optimize the build time using multi-threading. On the upside,
the build becomes more reproducible and the build result tends to be
smaller thanks to better optimization opportunities.

This is needed on gcc 11 for now build kodama: without it, the code
grows beyond the bounds it needs to fit in (that gcc 8 manages to hold,
so a regression).

BUG=none
BRANCH=none
TEST=one class of error less with gcc 11

Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Change-Id: I86b2c70cb5a2e85024630e7217cf8bd24d349910