TCPMv2: Ignore PR swap requests when PR transitioning
The charge manager may request a PR swap when it is no longer charging
from a port that is DRP and configured as sink.  This can prevent us
from depleting a tablet when there's a better charger available, for
example.

However, when the DRP partner has already initiated a PR swap this may
lead to the chromebook inadvertently becoming the sink of the connection
again.  Ignore PR swap requests while we're in a state where our PR role
doesn't match what it's expected to be shortly in order to prevent this.

BRANCH=None
BUG=b:174334068
TEST=on storo, connect guybrush and enable charging alternately from
both DUTs.  Then, when charging from the other DUT, connect a charger
and ensure that storo swaps to sourcing the other system.

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I1908f4ee726d49ca89f6f0c8f75bdfdf0fb12a47