util: Add function to convert a base3 number to binary first
Add an util function to convert a base3 number, which is in a normal
ternary number system, to a non-standard ternary number system where
the first 2^n natural numbers are represented as they would be in a
binary system (without any Z digits) and the following 3^n-2^n numbers
use the remaining ternary representations in the normal ternary system
order (skipping the values that were already used up).

This function is useful for converting BOARd ID, which is initially
used a binary and later decided to switch to tri-state after some
revisions have already been built.

BRANCH=Trogdor
BUG=b:190250108
TEST=make runhosttests

Change-Id: I853a04f3b28eb54c61855251dc5232c7e6994fef
Signed-off-by: Wai-Hong Tam <waihong@google.com>