util: Add function to convert a base3 number to binary first
Add an util function to convert a base3 number, which is in a normal
ternary number system, to a non-standard ternary number system where
the first 2^n natural numbers are represented as they would be in a
binary system (without any Z digits).

This function is used when running out of the BOARD IDs. The BOARD
IDs is usually first in the binary interpretation, 0 or 1. When
running out, we can switch to use the ternary interpretation, 0, 1,
or 2 (tri-state). This function helps the conversion without
affecting the original binary interpretation.

BRANCH=Trogdor
BUG=b:190250108
TEST=make runhosttests

Change-Id: I853a04f3b28eb54c61855251dc5232c7e6994fef
Signed-off-by: Wai-Hong Tam <waihong@google.com>