lazor: Interpret the BOARD IDs in the binary-first base3 system
Check the tri-state of the BOARD ID GPIOs. It gets a base3 number.
Convert the number to the binary-first base3 system. So it won't
affect the original binary BOARD ID assignment.

If all the GPIOs are not tri-state, it ends up to be a number in
binary interpretation.

BRANCH=Trogdor
BUG=b:190250108
TEST=Build the image and check the BOARD IDs.

Change-Id: Ibd938fbc05977df1eb7d8d27460beab4cd0f7bf7
Signed-off-by: Wai-Hong Tam <waihong@google.com>