cbi: Clean up cbi_create()
It doesn't seem to be helpful for cbi_create() to return anything.
So make it return void.

Besides, we don't need to define a local variable in cbi_create() for
accessing the head of the CBI cache, as there is already a static
cbi_header pointer with the scope of the whole file for this purpose.

BRANCH=None
BUG=b:186264627
TEST=make buildall -j

Signed-off-by: Philip Chen <philipchen@google.com>
Change-Id: Ie3e8bc06fa11e7071124e5aa3916e4c590d4a02d