bep/fpc_private: Enable sensor when FP private driver is unavailable
After EC reboot, FP_RST_ODL is set to 0, so FP sensor is not working.
FP private driver is responsible for enabling sensor using
fpc_sensor_spi_reset() function however private driver is not available
when building unit tests, so in this case we need to enable sensor.

This is part of fpsensor_hw unit test fix.

BUG=b:170432597
BRANCH=none
TEST='make BOARD=bloonchipper'

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I0e99a8f683258ecce58b29de82defc047b5d468b