ccd_config: remove UpdateNoTPMWipe
UpdateNoTPMWipe did not do anything. Replace it with 'Unused' in the
capabilities list.

We can't completely remove the capability because of how cr50 uses
the cap_info list. Every item from 0 to CCD_CAP_COUNT needs to have a
name and a default value. Renaming it makes it clear the capability
isn't used and keeps the relationship between capabilities and their
value in the capabilities bitmask.

BUG=b:182482706
TEST=none

Change-Id: Ie1e6674f414789ae52508b89e6f95d0a122cbd6f
Signed-off-by: Mary Ruthven <mruthven@chromium.org>