gingerbread: Account for port C1 1.5A current limit
This CL adds an override function for
typec_get_default_current_limit_rp() for gingerbread. This allows port
C0 to have a 3.0A current limit as desired, and still limit C1 to
1.5A.

A new macro BOARD_C1_1A5_LIMIT was added to allow this feature to be
used for honeybuns variants. This macro is also used to adjust the
USB-PD SRC CAP current limit to 1.5A

BUG=b:191793195
BRANCH=quiche
TEST=verfied on Gingerbread that Rp = 3.0A is selected for C0 and Rp =
1.5A is selected for port C1.

C0:
> ucpd info
        cc1      = Rp
        cc2      = Open
        Rp       = Rp_3.0
        cc1_v    = 2
        cc2_v    = 0
        rx_en    = 1
        pol      = 0

C1:
> tcpc 1 dump

  ROLE_CTRL                     (0x1a) =   0x15

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I2853ddcbc7b1d76320e480bd4c11ad555dd52bad