power: Reorder the power_chipset_init and the interrupt enable
The previous order is:
 * power_chipset_init()
 * enabling/disabling the power signal interrupts

There is a need that power_chipset_init() modifies the power signal
interrupt enabling/disabling, like crrev.com/c/2977573. The
power_chipset_init() checks if AP is ON or OFF; when AP is OFF, a
power signal interrupt is disabled (default) to avoid the floating
interrupt storm; but when AP is ON, the power signal interrupt is
enabled. But the next action resets the interrupt to default.

This CL reorders it to:
 * enabling/disabling the power signal interrupts
 * power_chipset_init()

So the interrupt enable/disable status won't be overridden.

BRANCH=None
BUG=b:185551931
TEST=Tested on Coachz, reboot EC, sysjump to RW, perform `apreset` which
executes a warm AP reboot.

Change-Id: I1740fed1d37a312ac822b2a91e9ebd5502756a1f
Signed-off-by: Wai-Hong Tam <waihong@google.com>