Guybrush: Switch to TCPC sourcing detection
Both PPC drivers used by guybrush report Vbus sourcing through a set of
cached flags.  However, with FRS in the picture their flags will no
longer accurately reflect whether we're sourcing.  The TCPC however will
correctly report that we're sourcing in its power control register.
This should also be a more reliable source of information when the
system is resetting or sysjumping.

It no longer seems necessary to confirm that a port is not sourcing
before performing a power supply reset, so this has been removed as
well.

BRANCH=None
BUG=b:183586640
TEST=on guybrush with FRS enabled, perform a failed FR swap and see
that Type-C ErrorRecovery correctly turns off Vbus.  Perform a
successful FR swap and ensure we don't have a lingering Vbus supplier in
the charge manager.  With no battery, boot from both ports with only a
charger plugged in.

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I7833f52dc84e2d5b097c4deb6acf7f2c29b8d48a