battery: Set host's low battery shutdown SoC to 3%
Currently, the host's low battery shutdown SoC is 2%. This is the
same as EC's low battery shutdown threshold. The EC waits for 30 secs
before it triggers the low battery shutdown and powerd reads the SoC
every 30 secs. Thus, in most cases powerd has a chance to shut down
the system gracefully.

This patch increases the host's shutdown SoC to 3%. This will further
gurantee that powerd will be given enough time to do everything for
a proper shutdown. It also avoids deeply discharging the battery,
which is bad for the battery health.

BUG=b:191837893
BRANCH=None
TEST=Altas using battfake EC command.

Change-Id: I3ab23205b400a1a326a60b8f9501611c027183b2
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>