trogdor: Enable battery fuel gauge
Do the same thing as CL:2987915, for Trogdor.

Enable the battery fuel gauge and add the battery info as an example,
such that the follower boards can just add the new battery info.
I pick a battery from the Lazor board. As long as it is a 2-cell
battery, it does not matter too much.

BRANCH=None
BUG=b:184071830
TEST=Built the trogdor image successfully and booted to AP.

Change-Id: Ic861a46148d238ba1dc015cd7b907321b484deca
Signed-off-by: Wai-Hong Tam <waihong@google.com>