zephyr: trogdor: Create Zephyr image for trogdor
Trogdor is a reference design, which doesn't have Zephyr support yet.
Lazor is a shipping design of the Trogdor family, which already have
Zephyr support. This CL creates the Trogdor Zephyr image and modifies
the hardware difference.

The CL first copies the directory lazor/ to trogdor/.

And do the following changes:
 * Rename all the board term lazor -> trogdor
 * Update the GPIOs (recheck the result of gpios_to_zephyr_dts)
 * Update the GPIO interrupts
 * Update the battery info
 * Update the sensor info
 * Do not include sku.c (trogdor doesn't have it yet)

BRANCH=None
BUG=b:184071830
TEST=Built the Zephyr image. Booted it on an existing Trogdor board.

Change-Id: I3b603284c719c119bb05460e7d4386d2c2eabda4
Signed-off-by: Wai-Hong Tam <waihong@google.com>