zephyr: implement ro/rw packing for hayato
The "npcx" packer in zmake was hard-coded for Nuvoton devices.
Restructure into a new packer, BinmanPacker, which makes no Nuvoton
assumptions, and subclass it for the NpcxPacker.

Note: there's some file size check here which was coded for Nuvoton
only.  I think it's not actually necessary since binman already checks
to see if the region overflowed, and will error if it did.  The best
thing to do here is to keep the existing behavior and evaluate if that
Nuvoton-only size check can just be removed from zmake.

BUG=b:192401039
BRANCH=none
TEST=zmake testall
TEST=inspected fmap of hayato image, note I don't have a device to
     test on right now

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ic45a030d0dd02451834161305f048487f7b273e4