EFS2: Jump to RW if system_is_manual_recovery is false
Currently, EFS2 tries to jump to RW if it's in RO on shutdown.
This hook was added to address the case when the system is shut down
from recovery screen by a power button press.

This change limits the condition to recovery mode only. This will
allow a system running on RO (for some reason) to stay in RO.

BUG=b:192463842, b:188242794
BRANCH=None
TEST=Storo. Verify EFS2 initiates a jump after shutdown from recovery
     screen.

Change-Id: I95148486c211d4a69d1b42ed94b9f641f82fc24c
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>