dooly: change lux calculations to avoid overflow
Previously there were several places where integer values could
overflow, such as converting 16-bit unsigned integers in the range [0,
64K] to fp_t values, which are are limited to the range [-32K, 32K].
Many multiplications also had the ability to produce results that
exceeded the range of fp_t. In practice this could cause overflows when
holding a phone flashlight in front of the sensor.

This rewrites most fixed point calculations with manual, but more
precise calculations on 64 bit types and keeps track of the bounds on
all values. Ideally we would use a consistent 64-bit fixed-point library
instead of approximating one, but this is a workable comprimise for a
relatevly short amount of code.

Also refactors a bit to reduce duplicated code.

BUG=b:179960346
BRANCH=none
TEST=builds

Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Change-Id: I52c37414b5205beaefa2111b1737b0eb22b7235a