bloonchipper: Reintroduce sleep lines to RO
Change https://crrev.com/c/2673909 removed the sleep lines from RO
in an effort to minimize the RO complexity.
Most notably, this isolated the deep-sleep/low-power-idle
logic to RW only.

Unfortunately, the sleep lines also control whether the SPI host interface
is listening, which allows it to ignore spurious communication.

It seems safer to reinstate the the sleep line with low power idle
active and directly disable CONFIG_LOW_POWER_IDLE in subsequent CL.

We reinstate the sleep line gpio logic from the following:
https://crrev.com/c5545464431669029f42829d542fa491d767ee5f/board/hatch_fp/board.c
This is the parent commit to the CL that refactors the sleep lines.

BRANCH=none
BUG=b:178746753
TEST=none

Signed-off-by: Craig Hesling <hesling@chromium.org>
Change-Id: I503cb3b62740300b265a4ddb165e29d9e36727fd