zephyr: volteer: Add CHRE project
Add skeleton code for the CHRE subsystem. This is the bare minimum needed
to build at least 1 CHRE file. More will be added later. Note that this
currently only builds with the Zephyr toolchain (see the TEST tag for
build details). Once coreboot-sdk migrates to gcc11 we can build with
the default toolchain in zmake.yaml.

BRANCH=none
BUG=b:191368705
TEST=First append chre.conf to prj.conf, then:
     $ zmake configure \
         --toolchain zephyr \
         -b zephyr/projects/volteer/volteer

Cq-Depend: chromium:3018813
Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I165a20c38a9f9e5a6ebe222c172c9e7f8a74f7a3