careena: Deassert PROCHOT on sysjump
The GPIO was an input in old RO images. On sysjump to new RW, the
direction is changed to output but the level is not set, which
results in the output driving low, which asserts PROCHOT incorrectly.

BUG=chromium:1226694 b:185082703
BRANCH=grunt
TEST=check PROCHOT is high after sysjump from old RO

Signed-off-by: Edward Hill <ecgh@chromium.org>
Change-Id: I81e6237d2643c58fa15672987b6ac815cbcd8211
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3016409
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
(cherry picked from commit e53eff67bebfdddcfd6ebb8d6ab468509fa340fc)