nocturne_fp/board_rw: Explicitly reset FP_RST_ODL to default on RW init
FP_RST_ODL pin is only defined in gpio_rw.inc, so this pin won't be
initialized to 1 by RO. Also, RW won't set this pin because sysjump to
RW is a warm reset actually and we are not setting GPIO pins to their
defaults after the warm reset to avoid undesired effects (see
gpio_pre_init() in chip/stm32/gpio.c).

As a result fpsensor_hw hardware unit test was not working, because
FPMCU was keeping sensor in reset.

Build with FP private driver is working because the driver performs
sensor reset during initialization.

BUG=b:170432597
BRANCH=none
TEST=`make BOARD=dartmonkey`

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Ia62340585da915e7d9f368107d57d3fb74cb831e