zmake: lazily create the logging thread
The logging thread was crashing the test_zmake.py logic that is using
the TestCase class. It seems that pytest was triggering the logging
thread to die (since it's set up as a daemon) when one TestCase finished
running. This caused the next TestCase to run into the .start() call
which crashed since we can't start the same thread multiple times.

BRANCH=none
BUG=none
TEST=zmake testall
TEST=pytest zephyr/zmake/tests

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I389f483a1d30d4ac21aaf547cd84c8adb9d8aa24