TCPMv2: Register Get_Source_Cap failures
When an attempt to gather source capabilities fails, we should avoid
probing for them again by internally tracking the number of capabilities
as -1.  This saves us traffic on resets with partners that have no
source capabilities, and also prevents looping with buggy partners who
reply with an unexpected message.

BRANCH=None
BUG=b:191229962
TEST=on guybrush, ensure we only query a sink-only dongle once

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Iea56619d7aca5df7d1b38d8f0d6a69cac4a825a9